Assets/Scripts/Tests/LineRendererCharactersTest.cs
using System.Collections;
using System.Collections.Generic;
using NUnit.Framework;
using UnityEngine;
using UnityEngine.TestTools;

using MyGame;
public class LineRendererCharactersTest
{
    private LineRendererCharacters lineRendererCharacters;
    private GameObject testGameObject;

    [SetUp]
    public void Setup()
    {
        // Create a new GameObject and add the LineRendererCharacters component
        //testGameObject = new GameObject("TestGameObject");
        //lineRendererCharacters = testGameObject.AddComponent<LineRendererCharacters>();
    }

    [UnityTest]
    public IEnumerator TestCharacterCreation()
    {
        // Define the character and its expected number of line segments
        //char testCharacter = 'A';
        //int expectedLineSegments = 5;

        // Create the test character using the LineRendererCharacters script
        //GameObject createdCharacter = lineRendererCharacters.CreateCharacter(testCharacter, Vector3.zero);

        // Check if the created character is not null
        //Assert.IsNotNull(createdCharacter);

        // Check if the created character has a LineRenderer component
        //LineRenderer lineRenderer = createdCharacter.GetComponent<LineRenderer>();
        //Assert.IsNotNull(lineRenderer);

        // Check if the created character has the expected number of line segments
        //Assert.AreEqual(expectedLineSegments, lineRenderer.positionCount);

        // Wait for one frame to pass
        yield return null;
    }

    [TearDown]
    public void TearDown()
    {
        // Clean up the test objects
        //Object.Destroy(testGameObject);
    }
}
